% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjMat2Graph}
\alias{adjMat2Graph}
\title{Convert an adjacency matrix into a graph}
\usage{
adjMat2Graph(adj, nodes, directed = FALSE)
}
\arguments{
\item{adj}{The adjacency matrix, NxN}

\item{nodes}{The names of the nodes, length N}

\item{directed}{TRUE if the graph should be directed, default FALSE}
}
\value{
The graph representation of the adjacency list
}
\description{
Convert an adjacency matrix into a graph
}
\examples{
mat <- matrix(sample(c(0,1), 16, replace=TRUE), nrow=4)
nodes <- c("X1", "X2", "X3", "X4")
g <- rCausalMGM::adjMat2Graph(mat, nodes, directed=TRUE)
}
