% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{allMetrics}
\alias{allMetrics}
\title{Combined graph recovery metrics}
\usage{
allMetrics(estimate, groundTruth, groundTruthDAG = NULL)
}
\arguments{
\item{estimate}{An estimated graph object}

\item{groundTruth}{A ground truth graph object of the same type as the estimated graph object}

\item{groundTruthDAG}{A ground truth graph object containing the true causal DAG. Only necessary for calculating the or precision, recall, F1, and MCC for partial ancestral graphs (PAGs)}
}
\value{
The orientation precision, recall, F1, and MCC, between the two graph objects
}
\description{
Calculate the SHD, precision, recall, F1, and Matthew's Correlation Coefficient (MCC) for the adjacencies and orientations of an estimated graph compared to the ground truth. This is the concatenated output of the SHD, adjacency PR metrics, and the orientation PR metrics.
}
\examples{
sim <- simRandomDAG(200, 25)
g <- pcStable(sim$data)
allMetrics(g, cpdag(sim$graph))
}
