% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap}
\alias{bootstrap}
\title{Runs bootstrapping for a selected causal discovery algorithm on the dataset.}
\usage{
bootstrap(
  df,
  algorithm = as.character(c("mgm-pc50", "mgm", "pc", "cpc", "pcm", "pc50", "fci",
    "cfci", "fcim", "mgm-pc", "mgm-cpc", "mgm-pcm", "mgm-fci", "mgm-cfci", "mgm-fcim",
    "mgm-fci50")),
  ensemble = as.character(c("highest", "majority")),
  lambda = as.numeric(c(0.2, 0.2, 0.2)),
  alpha = 0.05,
  numBoots = 20L,
  maxDiscrete = 5L,
  threads = -1L,
  verbose = FALSE
)
}
\arguments{
\item{df}{The dataframe}

\item{algorithm}{string indicating the name of the causal discovery algorithm to bootstrap. Causal discovery algorithms can be run alone or with mgm to learn an initial graph. Options include "mgm", "pc", "cpc", "pcm", "pc50", "fci", "cfci", "fcim", "mgm-pc", "mgm-cpc", "mgm-pcm", "mgm-pc50", "mgm-fci", "mgm-cfci", "mgm-fcim", "mgm-fci50." The default value is set to "mgm-pc50."}

\item{ensemble}{Method for construncting an ensemble graph from bootstrapped graphs. Options include "highest", which orients edges according to the orientation with the highest bootstrap probability, or "majority", which only orients edges if they have an orientation with a bootstrap probability > 0.5. Otherwise, the adjacency is included but the edge is left unoreineted. The default value is "highest."}

\item{lambda}{A vector of three lambda values - the first for continuous-continuous interaction, the second for continuous-discrete, and the third for discrete-discrete. Defaults to c(0.2, 0.2, 0.2). If a single value is provided, all three values in the vector will be set to that value.}

\item{alpha}{The p value below which results are considered significant. Defaults to 0.05.}

\item{numBoots}{The number of bootstrap samples to run. Defaults to 20.}

\item{maxDiscrete}{The maximum number of unique values a variable can have before being considered continuous. Defaults to 5}

\item{threads}{The number of consumer threads to create during multi-threaded steps. If -1, defaults to number of availible processors.}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated search graph with a table of edge stabilities
}
\description{
Runs bootstrapping for a selected causal discovery algorithm on the dataset.
}
\examples{
data("data.n100.p25")
g.boot <- rCausalMGM::bootstrap(data.n100.p25)
}
