% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpcStable}
\alias{cpcStable}
\title{Runs the causal algorithm CPC-Stable on a dataset}
\usage{
cpcStable(
  df,
  initialGraph = NULL,
  alpha = 0.05,
  threads = -1L,
  fdr = FALSE,
  rank = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{The dataframe}

\item{initialGraph}{An initial undirected graph to use as a starting point. If NULL, a full graph will be used. Defaults to NULL.}

\item{alpha}{The p value below which results are considered significant. Defaults to 0.05.}

\item{threads}{The number of consumer threads to create during multi-threaded steps. If -1, defaults to number of availible processors.}

\item{fdr}{Whether or not to run with FDR correction for the adjacencies.}

\item{rank}{Whether or not to use rank-based associations as opposed to linear}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated search graph
}
\description{
Runs the causal algorithm CPC-Stable on a dataset
}
\examples{
data("data.n100.p25")
ig <- rCausalMGM::mgm(data.n100.p25)
g <- rCausalMGM::cpcStable(data.n100.p25, initialGraph = ig)
}
