% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createMoral}
\alias{createMoral}
\title{Create the moral graph for the input directed acyclic graph (DAG). The moral graph is the equivalent undirected representation corresponding to the input DAG. The MGM algorithm learns the undirected moral graph for a corresponding causal DAG, so assessments of structure recovery should be compared to the moral graph rather than the causal DAG.}
\usage{
createMoral(graph)
}
\arguments{
\item{graph}{The graph object used to generate the moral graph. Should be the ground-truth causal DAG}
}
\value{
The moral graph corresponding to the input DAG
}
\description{
Create the moral graph for the input directed acyclic graph (DAG). The moral graph is the equivalent undirected representation corresponding to the input DAG. The MGM algorithm learns the undirected moral graph for a corresponding causal DAG, so assessments of structure recovery should be compared to the moral graph rather than the causal DAG.
}
\examples{
data(dag_n10000_p10)
moral <- rCausalMGM::createMoral(dag_n10000_p10)
}
