% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{data.n100.p25}
\alias{data.n100.p25}
\title{A small simulated dataset with 25 variables (13 discrete, 12 continuous) and 100 samples.
Generated using the true graph graph.n100.p25.txt}
\description{
A small simulated dataset with 25 variables (13 discrete, 12 continuous) and 100 samples.
Generated using the true graph graph.n100.p25.txt
}
\author{
Tyler Lovelace \email{tyl15@pitt.edu}
}
\keyword{data}
