% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{data.n1000.p100}
\alias{data.n1000.p100}
\title{A large simulated dataset with 100 variables (50 discrete, 50 continuous) and 1000 samples.
Generated using the true graph graph.n1000.p100.txt}
\description{
A large simulated dataset with 100 variables (50 discrete, 50 continuous) and 1000 samples.
Generated using the true graph graph.n1000.p100.txt
}
\author{
Tyler Lovelace \email{tyl15@pitt.edu}
}
\keyword{data}
