% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{graphTable}
\alias{graphTable}
\title{A function to generate a data.frame for objects from graph
class. It incorporates adjacency and orientation frequency if
estimates of edge stability are available.}
\usage{
graphTable(graph, stabilities = NULL)
}
\arguments{
\item{graph}{The graph object}

\item{stabilities}{The stability data.frame from bootstrapping or
StEPS. If NULL, the stabilities entry of the graph object is
used. If that is also NULL, only edge interactions are
returned. The default is NULL}
}
\value{
A data.frame containing source, target, and interaction
    columns for each edge in the graph. If stabilities are
    available, then the adjFrequency and orientation
    frequencies (if applicable) are returned for each edge.
}
\description{
A function to generate a data.frame for objects from graph
class. It incorporates adjacency and orientation frequency if
estimates of edge stability are available.
}
