% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadGraph}
\alias{loadGraph}
\title{Load a graph from a ".txt" file}
\usage{
loadGraph(filename)
}
\arguments{
\item{filename}{The graph file}
}
\value{
The graph as a graph object, which can be passed into search functions
}
\description{
Load a graph from a ".txt" file
}
