% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgm}
\alias{mgm}
\title{Calculate the MGM graph on a dataset}
\usage{
mgm(
  df,
  lambda = as.numeric(c(0.2, 0.2, 0.2)),
  maxDiscrete = 5L,
  verbose = as.logical(c(0))
)
}
\arguments{
\item{df}{The dataframe}

\item{lambda}{A vector of three lambda values - the first for continuous-continuous interaction, the second for continuous-discrete, and the third for discrete-discrete. Defaults to c(0.2, 0.2, 0.2)}

\item{maxDiscrete}{The maximum number of unique values a variable can have before being considered continuous. Defaults to 5}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated MGM graph
}
\description{
Calculate the MGM graph on a dataset
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgm("data.n100.p25")
}
