% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgmPath}
\alias{mgmPath}
\title{Calculate the solution path for an MGM graph on a dataset}
\usage{
mgmPath(df, lambdas = NULL, nLambda = 30L, rank = FALSE, verbose = FALSE)
}
\arguments{
\item{df}{The dataframe}

\item{lambdas}{A range of lambda values used to calculate a solution path for MGM. If NULL, lambdas is set to nLambda logarithmically spaced values from 10*sqrt(log10(p)/n) to sqrt(log10(p)/n). Defaults to NULL.}

\item{nLambda}{The number of lambda values to fit an MGM for when lambdas is NULL}

\item{rank}{Whether or not to use rank-based associations as opposed to linear}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated MGM graph
}
\description{
Calculate the solution path for an MGM graph on a dataset
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgmPath(data.n100.p25)
}
