% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orientationSHD}
\alias{orientationSHD}
\title{Calculate the orientation Structural Hamming Distance (SHD) between two graphs. This only counts the incorrect edge endpoints for edges present in both graphs, and does not consider differences in the graph skeleton. Each different endpoint adds 0.5 to the orientation SHD (i.e. A o-> B vs. A --> B). Thus, a completely misoriented edge adds 1 to the orientation SHD (i.e. A o-o B vs. A <-- B).}
\usage{
orientationSHD(graph1, graph2)
}
\arguments{
\item{graph1}{A graph object}

\item{graph2}{A graph object}
}
\value{
The skeleton SHD btween the two graph objects
}
\description{
Calculate the orientation Structural Hamming Distance (SHD) between two graphs. This only counts the incorrect edge endpoints for edges present in both graphs, and does not consider differences in the graph skeleton. Each different endpoint adds 0.5 to the orientation SHD (i.e. A o-> B vs. A --> B). Thus, a completely misoriented edge adds 1 to the orientation SHD (i.e. A o-o B vs. A <-- B).
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgm(data.n100.p25)
rCausalMGM::printGraph(g)
}
