% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{plot.graph}
\alias{plot.graph}
\title{A plot override function for the graph class}
\usage{
\method{plot}{graph}(x, nodes = c(), nodeAttr = list(), edgeAttr = list(), ...)
}
\arguments{
\item{x}{The graph object}

\item{nodes}{A subset of nodes in the graph to plot. If only a
single node is supplied, then that node and its Markov blanket
will be plotted.}

\item{nodeAttr}{A list of options to modify graph nodes
(e.g. fontsize).}

\item{edgeAttr}{A list of options to modify graph edges.}

\item{...}{Additional plot arguments}
}
\description{
A plot override function for the graph class
}
