% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{plot.graphCV}
\alias{plot.graphCV}
\title{A plot override function for the graphCV class}
\usage{
\method{plot}{graphCV}(x, ...)
}
\arguments{
\item{x}{The graph object}

\item{...}{Additional plot arguments}
}
\value{
No return value. This function plots graph sparsity, quantified by the average Markov blanket size for causal graphs or the regularization parameter for undirected graphs, against -log(pseudo-likelihood), with lines indicating the selected models.
}
\description{
A plot override function for the graphCV class
}
