% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{plot.graphPath}
\alias{plot.graphPath}
\title{A plot override function for the graphPath class}
\usage{
\method{plot}{graphPath}(x, ...)
}
\arguments{
\item{x}{The graph object}

\item{...}{Additional plot arguments}
}
\value{
No return value. This function plots graph sparsity, quantified by the regularization parameter, against the AIC and BIC scores along a solution path, with lines indicating the selected models.
}
\description{
A plot override function for the graphPath class
}
