% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{plot.graphSTARS}
\alias{plot.graphSTARS}
\title{A plot override function for the graphSTARS class}
\usage{
\method{plot}{graphSTARS}(x, ...)
}
\arguments{
\item{x}{The graph object}

\item{...}{Additional plot arguments}
}
\value{
No return value. This function plots graph sparsity, quantified by the significance threshold alpha, against the average edge instability used for stability-based model selection, with a horizontal line indicating the instability threshold and a vertical line indicating the selected threshold.
}
\description{
A plot override function for the graphSTARS class
}
