% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{print.knowledge}
\alias{print.knowledge}
\title{A print override function for the knowledge class}
\usage{
\method{print}{knowledge}(x, ...)
}
\arguments{
\item{x}{The knowledge object}

\item{...}{Additional print arguments}
}
\value{
No return value, the function prints a summary of the knowledge object.
}
\description{
A print override function for the knowledge class
}
