% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SHD}
\alias{SHD}
\title{Calculate the Structural Hamming Distance (SHD) between two graphs. This is the sum of the skeleton SHD and the orientation SHD.}
\usage{
SHD(graph1, graph2)
}
\arguments{
\item{graph1}{A graph object}

\item{graph2}{A graph object}
}
\value{
The SHD btween the two graph objects
}
\description{
Calculate the Structural Hamming Distance (SHD) between two graphs. This is the sum of the skeleton SHD and the orientation SHD.
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgm(data.n100.p25)
rCausalMGM::printGraph(g)
}
