% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skeletonSHD}
\alias{skeletonSHD}
\title{Calculate the skeleton Structural Hamming Distance (SHD) between two graphs. This only counts the missing and added edges, and does not consider edge orientation}
\usage{
skeletonSHD(graph1, graph2)
}
\arguments{
\item{graph1}{A graph object}

\item{graph2}{A graph object}
}
\value{
The skeleton SHD btween the two graph objects
}
\description{
Calculate the skeleton Structural Hamming Distance (SHD) between two graphs. This only counts the missing and added edges, and does not consider edge orientation
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgm(data.n100.p25)
rCausalMGM::printGraph(g)
}
