% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{steps}
\alias{steps}
\title{Calculate the optimal lambda values for the MGM algorithm and run the algorithm using those values. Optimal values are printed}
\usage{
steps(
  df,
  maxDiscrete = 5L,
  lambda = NULL,
  g = 0.05,
  numSub = 20L,
  leaveOneOut = as.logical(c(FALSE)),
  computeStabs = as.logical(c(FALSE)),
  verbose = as.logical(c(FALSE))
)
}
\arguments{
\item{df}{The dataframe}

\item{maxDiscrete}{The maximum number of unique values a variable can have before being considered continuous. Defaults to 5}

\item{lambda}{A vector of the lambda values to test. Defaults to a logspaced vector with 20 values ranging from 0.9 to 0.09 if n < p, or from 0.9 to 0.009 if n > p.}

\item{g}{The gamma parameter for STEPS. Defaults to 0.05}

\item{numSub}{The number of subsets to split the data into. Defaults to 20}

\item{leaveOneOut}{If TRUE, performs leave-one-out subsampling. Defaults to FALSE.}

\item{computeStabs}{If TRUE, stability values are calculated. Defaults to FALSE.}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated MGM graph
}
\description{
Calculate the optimal lambda values for the MGM algorithm and run the algorithm using those values. Optimal values are printed
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::steps(data.n100.p25)
}
