% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpcStable}
\alias{cpcStable}
\title{Runs the causal algorithm CPC Stable on a dataset}
\usage{
cpcStable(
  df,
  maxDiscrete = 5L,
  initialGraph = NULL,
  alpha = 0.05,
  verbose = as.logical(c(0))
)
}
\arguments{
\item{df}{The dataframe}

\item{maxDiscrete}{The maximum number of unique values a variable can have before being considered continuous. Defaults to 5}

\item{initialGraph}{The MGM graph to use as a starting point. If NULL, a full graph will be used. Defaults to NULL.}

\item{alpha}{The p value below which results are considered significant. Defaults to 0.05.}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated search graph
}
\description{
Runs the causal algorithm CPC Stable on a dataset
}
\examples{
df <- read.table("data/data.n1000.p25.txt", header=T)
ig <- rCausalMGM::mgm(df)
g <- rCausalMGM::cpcStable(df, initialGraph = ig)
}
