% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgmPath}
\alias{mgmPath}
\title{Calculate the solution path for an MGM graph on a dataset}
\usage{
mgmPath(df, lambdas = NULL, nLambda = 30L, maxDiscrete = 5L, verbose = FALSE)
}
\arguments{
\item{df}{The dataframe}

\item{lambdas}{A range of lambda values used to calculate a solution path for MGM. If NULL, lambdas is set to nLambda logarithmically spaced values from 10*sqrt(log10(p)/n) to sqrt(log10(p)/n). Defaults to NULL.}

\item{maxDiscrete}{The maximum number of unique values a variable can have before being considered continuous. Defaults to 5}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated MGM graph
}
\description{
Calculate the solution path for an MGM graph on a dataset
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgm(data.n100.p25)
}
