% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{plot.graphPath}
\alias{plot.graphPath}
\title{A plot override function for the graphCV class}
\usage{
\method{plot}{graphPath}(x, ...)
}
\arguments{
\item{x}{The graph object}

\item{...}{Additional plot arguments}
}
\description{
A plot override function for the graphCV class
}
