% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{plot.graphSTARS}
\alias{plot.graphSTARS}
\title{A plot override function for the graphSTARS class}
\usage{
\method{plot}{graphSTARS}(x, ...)
}
\arguments{
\item{x}{The graph object}

\item{...}{Additional plot arguments}
}
\description{
A plot override function for the graphSTARS class
}
