% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{plot.graphSTEPS}
\alias{plot.graphSTEPS}
\title{A plot override function for the graphSTEPS class}
\usage{
\method{plot}{graphSTEPS}(x, ...)
}
\arguments{
\item{x}{The graph object}

\item{...}{Additional plot arguments}
}
\description{
A plot override function for the graphSTEPS class
}
