% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printGraph}
\alias{printGraph}
\title{Display a graph object as text}
\usage{
printGraph(graph)
}
\arguments{
\item{graph}{The graph object}
}
\description{
Display a graph object as text
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgm(data.n100.p25)
rCausalMGM::printGraph(g)
}
