% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printGraph}
\alias{printGraph}
\title{Display a graph object as text}
\usage{
printGraph(graph, df)
}
\arguments{
\item{graph}{The graph object}

\item{df}{The dataframe containing the variables used by the graph}
}
\description{
Display a graph object as text
}
\examples{
df <- read.table("data/data.n1000.p25.txt", header=T)
g <- rCausalMGM::mgm(df)
rCausalMGM::printGraph(g, df)
}
