% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printGraph}
\alias{printGraph}
\title{Display a graph object as text.}
\usage{
printGraph(graph)
}
\arguments{
\item{graph}{The graph object}
}
\description{
Display a graph object as text. This is the same format as written in ".txt" save files.
}
\examples{
sim <- simRandomDAG(200, 25)
g <- mgm(sim$data)
printGraph(g)
}
