% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{saveGraph}
\alias{saveGraph}
\title{Save a graph to a file}
\usage{
saveGraph(list, filename)
}
\arguments{
\item{list}{The graph object}

\item{filename}{The graph file}
}
\description{
Save a graph to a file
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::mgm(data.n100.p25)
rCausalMGM::saveGraph(g, "graphs/mgm_graph.txt")
}
