% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{saveGraph}
\alias{saveGraph}
\title{Save a graph to a file. Supported file types are ".txt" and ".sif".}
\usage{
saveGraph(graph, filename)
}
\arguments{
\item{graph}{The graph object}

\item{filename}{The graph filename}
}
\value{
No return value. This function saves the full details of a graph object to a .txt file, including nodes, edges, the algorithm used to learn the model, and relevant hyperparameters. This format can then be read back into R with the loadGraph function.
}
\description{
Save a graph to a file. Supported file types are ".txt" and ".sif".
}
