% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{saveGraph}
\alias{saveGraph}
\title{Save a graph to a file. Supported file types are ".txt" and ".sif".}
\usage{
saveGraph(graph, filename)
}
\arguments{
\item{graph}{The graph object}

\item{filename}{The graph filename}
}
\description{
Save a graph to a file. Supported file types are ".txt" and ".sif".
}
