% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SHD}
\alias{SHD}
\title{Structural Hamming Distance (SHD)}
\usage{
SHD(graph1, graph2)
}
\arguments{
\item{graph1}{A graph object}

\item{graph2}{A graph object}
}
\value{
The SHD btween the two graph objects
}
\description{
Calculate the Structural Hamming Distance (SHD) between two graphs.
}
\examples{
sim <- simRandomDAG(200, 25)
g <- pcStable(sim$data)
SHD(g, cpdag(sim$graph))
}
