% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{steps}
\alias{steps}
\title{Calculate the optimal lambda values for the MGM algorithm and run the algorithm using those values. Optimal values are printed}
\usage{
steps(
  df,
  maxDiscrete = 5L,
  lambda = as.numeric(c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, 0.45, 0.5, 0.55,
    0.6, 0.65, 0.7, 0.75, 0.8, 0.85)),
  g = 0.05,
  numSub = 20L,
  leaveOneOut = as.logical(c(0)),
  computeStabs = as.logical(c(0)),
  verbose = as.logical(c(0))
)
}
\arguments{
\item{df}{The dataframe}

\item{maxDiscrete}{The maximum number of unique values a variable can have before being considered continuous. Defaults to 5}

\item{lambda}{A vector of the lambda values to test. Defaults to c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, 0.45, 0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85)}

\item{g}{The gamma parameter for STEPS. Defaults to 0.05}

\item{numSub}{The number of subsets to split the data into. Defaults to 20}

\item{leaveOneOut}{If TRUE, performs leave-one-out subsampling. Defaults to FALSE.}

\item{computeStabs}{If TRUE, stability values are calculated. Defaults to FALSE.}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated MGM graph
}
\description{
Calculate the optimal lambda values for the MGM algorithm and run the algorithm using those values. Optimal values are printed
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::steps(data.n100.p25)
}
