% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{steps}
\alias{steps}
\title{Calculates the optimal lambda values for the MGM algorithm using StEPS and run the algorithm using those values. Optimal values are printed}
\usage{
steps(
  df,
  lambdas = NULL,
  nLambda = 20L,
  g = 0.05,
  numSub = 20L,
  subSize = -1L,
  leaveOneOut = FALSE,
  computeStabs = FALSE,
  threads = -1L,
  rank = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{The dataframe}

\item{lambdas}{A range of lambda values assessed for stability by the StEPS algorithm. If NULL, lambdas is set to nLambda logarithmically spaced values from 10*sqrt(log10(p)/n) to sqrt(log10(p)/n). Defaults to NULL.}

\item{nLambda}{The number of lambda values to fit an MGM for when lambdas is NULL}

\item{g}{The gamma parameter for STEPS. Defaults to 0.05}

\item{numSub}{The number of subsets to split the data into. Defaults to 20}

\item{subSize}{The size of the subsamples used for STEPS. If the value is -1, the size of the subsamples is set to floor(10*sqrt(n)). If the value is in the range (0,1), the size of the subsamples is set to floor(subSize * n). Otherwise, if subSize is in the range [1,n), the size of the subsamples is set to subSize. Defaults to -1.}

\item{leaveOneOut}{If TRUE, performs leave-one-out subsampling. Defaults to FALSE.}

\item{computeStabs}{If TRUE, stability values are calculated. Defaults to FALSE.}

\item{threads}{The number of consumer threads to create during multi-threaded steps. If -1, defaults to number of availible processors.}

\item{rank}{Whether or not to use rank-based associations as opposed to linear}

\item{verbose}{Whether or not to output additional information. Defaults to FALSE.}
}
\value{
The calculated MGM graph
}
\description{
Calculates the optimal lambda values for the MGM algorithm using StEPS and run the algorithm using those values. Optimal values are printed
}
\examples{
data("data.n100.p25")
g <- rCausalMGM::steps(data.n100.p25)
}
